Fixing issue with sticker rendition on android
I think the issue was with the use of implicit UV coordinates on the the triangle used for rendition.

What we were doing is drawing a triangle with the following relationship to the screen.
│\
│█\

And I think Android is screwing the coordinates up. So I've changed it to make it generate the UVs as a part of screen rendition.